+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Added
 
 - button to create default tasks
+- empty places in execution graph, when no tasks were executed
 
 # 1.0.5